dockerize the application
All files required for dockerizing are in this directory. However the directory structure is not correct. The correct directory structure is,
Dockerfile
start.sh
---------------------->api/
                             ----------------->app.py, requirements.txt, model/
To build the african fabric image file,
>>docker build -t  <img_name>  .
To run,
>>docker run -it --name <img_name> -p 5000:5000 --rm <img_name>

The GET and POST requests work with chrome browser. For the GET request open,
https://localhost:5000
For the POST request,
https://localhost:5000/predict
This should display a 8x8 grid of fabric images. Repeatedly hitting enter on the address bar should display a different set of images.